+++ example/pom.xml
@@ -58,5 +58,9 @@
             <artifactId>rocketmq-acl</artifactId>
             <version>4.6.1-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>rocketmq-tools</artifactId>
+        </dependency>
     </dependencies>
 </project>
