--- example/src/main/java/org/apache/rocketmq/example/tracemessage/TraceInfoConsumer.java
@@ -1,71 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.rocketmq.example.tracemessage;
-
-import org.apache.rocketmq.client.consumer.DefaultMQPushConsumer;
-import org.apache.rocketmq.client.consumer.listener.ConsumeConcurrentlyContext;
-import org.apache.rocketmq.client.consumer.listener.ConsumeConcurrentlyStatus;
-import org.apache.rocketmq.client.consumer.listener.MessageListenerConcurrently;
-import org.apache.rocketmq.client.exception.MQClientException;
-import org.apache.rocketmq.client.trace.TraceBean;
-import org.apache.rocketmq.client.trace.TraceContext;
-import org.apache.rocketmq.client.trace.TraceDataEncoder;
-import org.apache.rocketmq.common.MixAll;
-import org.apache.rocketmq.common.message.MessageExt;
-
-import java.util.List;
-
-/**
- * In this example we show you how to use message trace.
- * When the enableMsgTrace flag of producer/consumer is true,
- * trace data will be sent to RMQ_SYS_TRACE_TOPIC or customized topic(if provided),
- * so we can get message trace data by subscribe the topic.
- */
-public class TraceInfoConsumer {
-    public static void main(String[] args) throws MQClientException {
-        DefaultMQPushConsumer consumer = new DefaultMQPushConsumer("C_TRACE");
-        // Use the default message trace topic
-        consumer.subscribe(MixAll.RMQ_SYS_TRACE_TOPIC, "*");
-        consumer.registerMessageListener(new MessageListenerConcurrently() {
-            @Override
-            public ConsumeConcurrentlyStatus consumeMessage(List<MessageExt> msgs, ConsumeConcurrentlyContext context) {
-                for (int i = 0; i < msgs.size(); i++) {
-                    // Decode trace data
-                    List<TraceContext> traceContexts = TraceDataEncoder.decoderFromTraceDataString(new String(msgs.get(i).getBody()));
-                    for (TraceContext tc: traceContexts) {
-                        for (TraceBean tb: tc.getTraceBeans()) {
-                            StringBuilder sb = new StringBuilder();
-                            sb.append(tb.getClientHost()).append(" ")
-                                    .append(tc.getGroupName()).append(" ")
-                                    .append(tc.getTimeStamp()).append(" ")
-                                    .append(tc.getTraceType()).append(" ")
-                                    .append(tb.getTopic()).append(" ")
-                                    .append(tb.getMsgId()).append(" ")
-                                    .append(tb.getMsgType()).append(" ");
-                            sb.append("%n");
-                            System.out.printf(sb.toString());
-                        }
-                    }
-                }
-                return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
-            }
-        });
-        consumer.start();
-        System.out.printf("Consumer Started.%n");
-    }
-}
